refactor: 컨벤션 통일 #160
