+++ src/main/java/com/muud/auth/controller/AuthController.java
@@ -11,7 +11,6 @@ import com.muud.global.error.ExceptionType;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -21,6 +20,7 @@ import java.util.Map;
 @RestController
 @RequiredArgsConstructor
 public class AuthController {
+
     private final AuthService authService;
     private final KakaoService kakaoService;
     @Value("${admin-code}")
