+++ src/main/java/com/muud/diary/domain/Diary.java
@@ -3,7 +3,7 @@ package com.muud.diary.domain;
 import com.muud.bookmark.domain.Bookmark;
 import com.muud.emotion.domain.Emotion;
 import com.muud.global.common.BaseEntity;
-import com.muud.playlist.entity.PlayList;
+import com.muud.playlist.domain.PlayList;
 import com.muud.user.entity.User;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
