+++ src/main/java/com/muud/diary/domain/dto/DiaryResponse.java
@@ -2,7 +2,7 @@ package com.muud.diary.domain.dto;
 
 import com.muud.diary.domain.Diary;
 import com.muud.emotion.domain.Emotion;
-import com.muud.playlist.dto.VideoDto;
+import com.muud.playlist.domain.dto.VideoDto;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
