+++ src/main/java/com/muud/playlist/controller/PlayListController.java
@@ -6,12 +6,11 @@ import com.muud.global.common.PageResponse;
 import com.muud.global.error.ApiException;
 import com.muud.global.error.ExceptionType;
 import com.muud.global.error.ResponseError;
-import com.muud.playlist.dto.PlayListRequest;
-import com.muud.playlist.dto.VideoDto;
+import com.muud.playlist.domain.dto.PlayListRequest;
+import com.muud.playlist.domain.dto.VideoDto;
 import com.muud.playlist.service.PlayListService;
 import com.muud.playlist.service.YoutubeDataService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
@@ -27,10 +26,10 @@ import static com.muud.auth.jwt.Auth.Role.ADMIN;
 @RestController
 @RequiredArgsConstructor
 public class PlayListController {
+
     private final PlayListService playListService;
     private final YoutubeDataService youtubeDataService;
-    @Value("${admin-code}")
-    private String ADMIN_CODE;
+
     @GetMapping("/playlists")
     public ResponseEntity<PageResponse> getPlayLists(@RequestParam(name = "emotion") Emotion emotion, @PageableDefault(size = 4) Pageable pageable){
         Page<VideoDto> videoDtoList = playListService.getPlayLists(emotion, pageable);
@@ -48,6 +47,7 @@ public class PlayListController {
             throw new ApiException(ExceptionType.SYSTEM_ERROR);
         }
     }
+
     @Auth(role = ADMIN)
     @PostMapping("/playlists")
     public ResponseEntity addPlayList(@RequestBody PlayListRequest playListRequestList){
@@ -56,12 +56,14 @@ public class PlayListController {
                 .body(Map.of("message", "success"));
 
     }
+
     @Auth(role = ADMIN)
     @DeleteMapping("/playlists/{playlistId}")
     public ResponseEntity deletePlayList(@PathVariable Long playlistId){
         playListService.removePlayList(playlistId);
         return ResponseEntity.noContent().build();
     }
+
     @ExceptionHandler(IllegalArgumentException.class)
     @ResponseStatus(HttpStatus.BAD_REQUEST)
     public ResponseEntity<ResponseError> handleInvalidEnumValueException(IllegalArgumentException ex) {
