+++ src/main/java/com/muud/playlist/repository/PlayListRepository.java
@@ -1,7 +1,7 @@
 package com.muud.playlist.repository;
 
 import com.muud.emotion.domain.Emotion;
-import com.muud.playlist.entity.PlayList;
+import com.muud.playlist.domain.PlayList;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -10,6 +10,7 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.Optional;
 
 public interface PlayListRepository extends JpaRepository<PlayList, Long> {
+
     @Query("SELECT p from PlayList p where p.emotion = :emotion order by rand()")
     Page<PlayList> findByEmotion(Emotion emotion, Pageable pageable);
