+++ src/main/java/com/muud/playlist/service/PlayListService.java
@@ -3,8 +3,8 @@ package com.muud.playlist.service;
 import com.muud.emotion.domain.Emotion;
 import com.muud.global.error.ApiException;
 import com.muud.global.error.ExceptionType;
-import com.muud.playlist.dto.VideoDto;
-import com.muud.playlist.entity.PlayList;
+import com.muud.playlist.domain.dto.VideoDto;
+import com.muud.playlist.domain.PlayList;
 import com.muud.playlist.repository.PlayListRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -14,7 +14,9 @@ import org.springframework.stereotype.Service;
 @Service
 @RequiredArgsConstructor
 public class PlayListService {
+
     private final PlayListRepository playListRepository;
+
     public Page<VideoDto> getPlayLists(Emotion emotion, Pageable pageable){
         return playListRepository.findByEmotion(emotion, pageable)
                .map(PlayList::toDto);
