+++ src/main/java/com/muud/report/service/ReportService.java
@@ -3,9 +3,9 @@ package com.muud.report.service;
 import com.muud.diary.domain.Diary;
 import com.muud.diary.repository.DiaryRepository;
 import com.muud.emotion.domain.Emotion;
-import com.muud.playlist.entity.PlayList;
+import com.muud.playlist.domain.PlayList;
 import com.muud.report.domain.dto.RankDto;
-import com.muud.playlist.dto.VideoDto;
+import com.muud.playlist.domain.dto.VideoDto;
 import com.muud.report.domain.dto.EmotionReport;
 import com.muud.report.domain.dto.PlaylistReport;
 import com.muud.report.domain.dto.ReportResponse;
