+++ src/main/java/com/muud/user/service/UserService.java
@@ -12,6 +12,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Service @Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class UserService {
+
     private final UserRepository userRepository;
 
     @Transactional
